--- a/quinn/src/lib.rs
+++ b/quinn/src/lib.rs
@@ -155,8 +155,8 @@ enum EndpointEvent {
     Transmit(proto::Transmit),
 }
 
-/// Maximum number of send/recv calls to make before moving on to other processing
+/// Maximum number of datagrams processed in send/recv calls to make before moving on to other processing
 ///
-/// This helps ensure we don't starve anything when the CPU is slower than the link. Value selected
-/// more or less arbitrarily.
-const IO_LOOP_BOUND: usize = 10;
+/// This helps ensure we don't starve anything when the CPU is slower than the link.
+/// Value is selected by picking a low number which didn't degrade throughput in benchmarks.
+const IO_LOOP_BOUND: usize = 160;