This contains 2 commits which in combination try to prevent the endpoint task from using too many executor resources while keeping performance or event improving on it.

Performance difference various a bit between the machinesI tested. On one there is no change, on another an increase in througput:

**Before:**
```
Sent 10737418240 bytes on 1 streams in 22.50s (455.04 MiB/s)
```

**After:**
```
Sent 10737418240 bytes on 1 streams in 21.92s (467.06 MiB/s)
```