@djc 2021-05-17
Sorry for not reviewing this sooner -- I was on holiday for two weeks. The reason I hadn't approved yet was because I was wondering about `handle_events()`. As I understand it, it is not currently limited by `IO_BOUND_LOOP`. Should we worry about a scenario where there are a lot of events to handle and we get stuck inside `handle_events()` for too long? I guess that's pretty much orthogonal to this PR, but still wanted to mention it.