@Matthias247 2021-05-17
> Should we worry about a scenario where there are a lot of events to handle and we get stuck inside handle_events() for too long? I guess that's pretty much orthogonal to this PR, but still wanted to mention it.

Yes, I think it also makes sense to limit those interactions. It however won't be an issue with a singlethreaded runtime since connections can't produce new events while the endpoint task is running, so it's guaranteed to terminate. The IO part was more risky since it gets events from the outside, and a peer sending data fast enough could prevent the endpoint task from ever yielding.

However for better performance with multithreaded runtimes it could make sense to limit the amount of events handled in `handle_events()`. As you said, it's kind of orthogonal.