@Ralith 2021-05-17
The expected rate of endpoint events is *much* lower than the expected rate of incoming/outgoing packets, FWIW. They're only raised during connection lifecycle events like handshaking, migration, and loss.