@Matthias247 2021-05-18
[`Endpoint::handle_events`](https://github.com/quinn-rs/quinn/blob/fa0ac1f85a70867c3ecb45ba1e778cd29db4d98f/quinn/src/endpoint.rs#L371-L401) does not only handle the endpoint related events. It receives all the outgoing messages from connections and enqueues them for transmission. This could theoretically lead to an unbounded `outgoing`  queue.

Another potential issue I noticed is in [`drive_send`](https://github.com/quinn-rs/quinn/blob/fa0ac1f85a70867c3ecb45ba1e778cd29db4d98f/quinn/src/endpoint.rs#L339-L344)  endpoint related datagrams are only enqueued if not enough other packets are enqueued. That could mean that in a busy endpoint those might never get sent. Probably that condition should just be dropped.